move CurrencyRepository to Repository ns
