+++ src/PrestaShopBundle/Resources/config/services/adapter/currency.yml
@@ -122,5 +122,5 @@ services:
       tags:
         - { name: 'tactician.handler', command: 'PrestaShop\PrestaShop\Core\Domain\Currency\Query\GetReferenceCurrency' }
 
-    prestashop.adapter.currency.currency_repository:
-        class: 'PrestaShop\PrestaShop\Adapter\Currency\CurrencyRepository'
+    prestashop.adapter.currency.repository.currency_repository:
+        class: 'PrestaShop\PrestaShop\Adapter\Currency\Repository\CurrencyRepository'
